Move slot module (non-breaking)
This moves the slot module out of types and into
the location indicated in commit 20b26a5.

src/types/slot_token.rs -> src/slot/mod.rs
src/types/mod.rs::{SlotFlags,TokenFlags} -> src/slot/flags.rs
